backend/requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic
pyyaml
numpy
pydantic_settings